+++ components/ArmyRoster/ArmyRoster.tsx
@@ -1,22 +1,35 @@
 import { type Army } from 'appdeptus/models'
-import React, { type ComponentProps, memo } from 'react'
-import { FlatList } from 'react-native'
+import React, {
+  type ComponentProps,
+  type ForwardedRef,
+  forwardRef,
+  memo
+} from 'react'
 import { VStack } from '../ui'
 import { UnitListItem } from '../UnitListItem'
+import Animated from 'react-native-reanimated'
+import { type FlatList } from 'react-native'
 
-type ArmyRosterProps = {
+type ArmyRosterProps = Partial<
+  Omit<ComponentProps<typeof Animated.FlatList<Army['roster'][number]>>, 'data'>
+> & {
   roster: Army['roster']
-  ListHeaderComponent?: ComponentProps<typeof FlatList>['ListHeaderComponent']
-
   invalidUnits?: Army['roster'][number]['id'][]
 }
 
-const ArmyRoster = ({
-  roster,
-  ListHeaderComponent,
-  invalidUnits = []
-}: ArmyRosterProps) => (
-  <FlatList
+const ArmyRoster = (
+  {
+    roster,
+    ListHeaderComponent,
+    invalidUnits = [],
+    ...flatListProps
+  }: ArmyRosterProps,
+  ref: ForwardedRef<FlatList<Army['roster'][number]>>
+) => (
+  <Animated.FlatList
+    {...flatListProps}
+    ref={ref}
+    scrollEventThrottle={16}
     data={roster}
     contentContainerClassName='pt-32'
     showsVerticalScrollIndicator={false}
@@ -45,4 +58,4 @@ const ArmyRoster = ({
   />
 )
 
-export default memo(ArmyRoster)
+export default memo(forwardRef(ArmyRoster))