+++ modules/armies/hooks/types.ts
@@ -1,0 +1,17 @@
+import { type MutableRefObject } from 'react'
+import { type PanGesture } from 'react-native-gesture-handler'
+import {
+  type SharedValue,
+  type ScrollHandlerProcessed
+} from 'react-native-reanimated'
+
+type ArmyRosterBackgroundAnimationObject = {
+  onScroll: ScrollHandlerProcessed
+  panGesture?: PanGesture
+  panGestureRef?: MutableRefObject<PanGesture | undefined>
+  panScrollValue?: SharedValue<number>
+  scrollValue: SharedValue<number>
+  scale: SharedValue<number>
+}
+
+export type { ArmyRosterBackgroundAnimationObject }